+++ src/Filament/Resources/SentMessageResource.php
@@ -24,7 +24,7 @@ class SentMessageResource extends Resource
 
     protected static ?string $navigationGroup = 'SMS Management';
 
-    protected static ?int $navigationSort = 1;
+    protected static ?int $navigationSort = 2;
 
     public static function form(Form $form): Form
     {